feature(#172): Extra turtlebot sensors
- Switch to eturtlebot3, because that is what I'm working with now

- Switch to shared library to reduce size

+++ include/cosm/convergence/angular_order.hpp
@@ -57,7 +57,7 @@ class angular_order final : public convergence_measure {
    * parameters and the current state of the swarm.
    */
   bool operator()(const std::vector<rmath::radians>& headings,
-                  RCPPSW_UNUSED uint n_threads) {
+                  RCPPSW_UNUSED size_t n_threads) {
     double y = 0.0;
     double x = 0.0;
 
@@ -73,4 +73,3 @@ class angular_order final : public convergence_measure {
 };
 
 NS_END(convergence, cosm);
-
