+++ include/cosm/ds/operations/cell2D_op.hpp
@@ -66,8 +66,8 @@ class cell2D_op {
 
   virtual ~cell2D_op(void) = default;
 
-  uint x(void) const { return m_coord.x(); }
-  uint y(void) const { return m_coord.y(); }
+  size_t x(void) const { return m_coord.x(); }
+  size_t y(void) const { return m_coord.y(); }
 
   const rmath::vector2z& coord(void) const { return m_coord; }
