+++ src/convergence/metrics/convergence_metrics_collector.cpp
@@ -52,27 +52,27 @@ void convergence_metrics_collector::collect(
   auto status = m.swarm_interactivity();
   m_data.interact.raw += std::get<0>(status);
   m_data.interact.norm += std::get<1>(status);
-  m_data.interact.converged += static_cast<uint>(std::get<2>(status));
+  m_data.interact.converged += static_cast<size_t>(std::get<2>(status));
 
   status = m.swarm_angular_order();
   m_data.order.raw += std::get<0>(status);
   m_data.order.norm += std::get<1>(status);
-  m_data.order.converged += static_cast<uint>(std::get<2>(status));
+  m_data.order.converged += static_cast<size_t>(std::get<2>(status));
 
   status = m.swarm_positional_entropy();
   m_data.pos_ent.raw += std::get<0>(status);
   m_data.pos_ent.norm += std::get<1>(status);
-  m_data.pos_ent.converged += static_cast<uint>(std::get<2>(status));
+  m_data.pos_ent.converged += static_cast<size_t>(std::get<2>(status));
 
   status = m.swarm_task_dist_entropy();
   m_data.tdist_ent.raw += std::get<0>(status);
   m_data.tdist_ent.norm += std::get<1>(status);
-  m_data.tdist_ent.converged += static_cast<uint>(std::get<2>(status));
+  m_data.tdist_ent.converged += static_cast<size_t>(std::get<2>(status));
 
   status = m.swarm_velocity();
   m_data.velocity.raw += std::get<0>(status);
   m_data.velocity.norm += std::get<1>(status);
-  m_data.velocity.converged += static_cast<uint>(std::get<2>(status));
+  m_data.velocity.converged += static_cast<size_t>(std::get<2>(status));
 } /* collect() */
 
 void convergence_metrics_collector::reset_after_interval(void) {
